Editted things mentiond in comments
This changes the order of somethings in the document to make it easier
to understand reading from the top to the bottom.
The requirements have a short form to make them easier to refer to.
Rationale has been added to the syntax section in the form of describing
how the syntax enables the requirements.
A brief discussion of fnmatch has been added
Thoughts on implementation have been removed.
A small section about dynamic remapping has been added.
This displays nicely in Jekyll