Add article about how ROS 2 integrates with DDS-Security
This is not a new design proposal. Rather, it is a description of the
way ROS 2 _currently_ integrates with DDS-Security (i.e. SROS 2). No new
ideas are contained here; this is meant to fill a hole in existing
design documentation and provide a good basis for future design of SROS
2, providing necessary context both for the designer as well as the
reviewer.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>